requirements.txt
facebook_sdk
Pillow
